+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
+# v2.5.0
+* Implement `Note(pitch_name::String; position = 0, velocity = 100, duration = 960, channel = 0)`
+* Implement `Notes(notes_string::String, tpq::Int = 960)`
 # v2.4.0
 * Implement `Base.empty!(::Notes)` and `Base.isempty(::Notes)`.
 # v2.3.0